Feature/follow 01

+++ backend/src/main/java/com/example/backend/identity/member/dto/MemberResponse.java
@@ -4,7 +4,7 @@ public record MemberResponse(
 	long id,
 	String username,
 	String profileUrl,
-	int followerCount,
-	int followingCount
+	long followerCount,
+	long followingCount
 ) {
 }
